add internal.{lo,hi}
This comes in a separate patch to a) show that it is very easy
to add the new primitives, and b) add some documentation.

So the new lo/hi primitives act as undocumented lowest/highest
but actually work.

Say, both

	process = lowest;
and
	process = internal.lo;

output

	output0[i0] = FAUSTFLOAT(-1.0f);

But, say,

	process = +(1) : lowest;

crashes the compiler, while

	process = +(1) : internal.lo;

correctly outputs

	output0[i0] = FAUSTFLOAT(0.0f);

(I guess the implementation of lowest/highest is not finished,
 with this patch this code can be removed).

//------------------------------------------------------------
Example:

	ge(x,y) = internal.lo(x) >= internal.hi(y);

returns 1 if the compiler can deduce at compile time that x >= y,
otherwise 0. So

	process = _,_ <: _,_, +(1),-(1) : ge,ge;

does not generate any code and outputs

	output0[i0] = FAUSTFLOAT(0);
	output1[i0] = FAUSTFLOAT(1);

Again, I don't think it is possible to implement this in faust.